updated state managemnt
